requirements.txt
anthropic>=0.82.0
python-dotenv>=1.0.0
